requirements.txt
fastapi
uvicorn
python-dotenv
openai
google-generativeai
pydantic